refactor(vapor): split operation generators
